+++ src/helpers/types/assert.ts
@@ -1,0 +1,6 @@
+/**
+ * Assertion function
+ */
+export function assert(value: unknown, message: string | Error): asserts value {
+  if (!value) throw message instanceof Error ? message : new Error(message);
+}
