+++ src/helpers/types/enumToArray.ts
@@ -1,3 +1,6 @@
+/**
+ * Converts a typescript enum to an array of it's values
+ */
 export const enumToArray = (_enum: any) => {
   const values = Object.values(_enum);
