+++ src/hooks/useENS.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "react-query";
 import { NetworkId } from "src/constants";
-import { queryAssertion } from "src/helpers/reactQueryUtil";
-import { nonNullable } from "src/helpers/typeUtil";
+import { queryAssertion } from "src/helpers/react-query/queryAssertion";
+import { nonNullable } from "src/helpers/types/nonNullable";
 
 import { useWeb3Context } from ".";
