+++ src/hooks/usePendingTransactions.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "react-query";
 import { NetworkId } from "src/constants";
-import { queryAssertion } from "src/helpers/reactQueryUtil";
-import { nonNullable } from "src/helpers/typeUtil";
+import { queryAssertion } from "src/helpers/react-query/queryAssertion";
+import { nonNullable } from "src/helpers/types/nonNullable";
 import { covalent } from "src/lib/covalent";
 import { CovalentTransaction } from "src/lib/covalent.types";
