+++ src/hooks/useStakingRebaseRate.ts
@@ -1,8 +1,9 @@
 import { useQuery } from "react-query";
 import { NetworkId } from "src/constants";
 import { SOHM_ADDRESSES, STAKING_ADDRESSES } from "src/constants/addresses";
-import { createDependentQuery, parseBigNumber } from "src/helpers";
-import { queryAssertion } from "src/helpers/reactQueryUtil";
+import { parseBigNumber } from "src/helpers";
+import { createDependentQuery } from "src/helpers/react-query/createDependentQuery";
+import { queryAssertion } from "src/helpers/react-query/queryAssertion";
 
 import { useStaticSohmContract, useStaticStakingContract } from "./useContract";
