+++ src/views/Stake/components/ExternalStakePools/hooks/useStakePoolTVL.ts
@@ -1,8 +1,9 @@
 import { useQuery } from "react-query";
 import { GOHM_ADDRESSES } from "src/constants/addresses";
-import { createDependentQuery, getTokenPrice, parseBigNumber } from "src/helpers";
-import { queryAssertion } from "src/helpers/reactQueryUtil";
-import { nonNullable } from "src/helpers/typeUtil";
+import { getTokenPrice, parseBigNumber } from "src/helpers";
+import { createDependentQuery } from "src/helpers/react-query/createDependentQuery";
+import { queryAssertion } from "src/helpers/react-query/queryAssertion";
+import { nonNullable } from "src/helpers/types/nonNullable";
 import { useStaticPairContract } from "src/hooks/useContract";
 import { useGohmPrice } from "src/hooks/usePrices";
 import { ExternalPool } from "src/lib/ExternalPool";
