+++ src/views/Stake/components/StakeArea/components/StakeActionArea/components/StakeNextRebaseAmount.tsx
@@ -2,7 +2,7 @@ import { t } from "@lingui/macro";
 import { DataRow } from "@olympusdao/component-library";
 import { BigNumber } from "ethers";
 import { convertGohmToOhm, formatNumber, parseBigNumber } from "src/helpers";
-import { nonNullable } from "src/helpers/typeUtil";
+import { nonNullable } from "src/helpers/types/nonNullable";
 import {
   useFuseBalance,
   useGohmBalance,
